+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Calendar Versioning](https://calver.org).
 - [CSE] Added support for the JSON-path content filter query in discovery requests. This allows to filter &lt;CIN> resources based on their content attribute.
 - [CSE] Added initial support for validating attributes based on the release version indicator (rvi) in requests. This allows to validate whether an attribute is allowed for a specific release version of the oneM2M standard.
 - [CSE] Added validation of *m2m:externalID* types, e.g. for M2M-EXT-ID attributes.
+- [CSE] Added support for the *m2m:triggerRecipientID* types, e.g. for triggerRecipientID attributes.
 - [MGMT] Added reloading of HTTP and WebSocket authentication credentials to the management API.
 - [MGMT] Added support for retrieving basic and auth token information via the management API.
 - [MGMT] Added return of attribute, resource type, and flexcontainer policies to the management API.